Refactor web server, geo data.
